+++ test/transform/resource/before/ArgumentIntFragment.java
@@ -1,0 +1,10 @@
+// ignore
+
+import android.app.Fragment;
+import android.os.Bundle;
+
+public class ArgumentIntFragment extends Fragment {
+	
+	@hrisey.Argument
+	private int myInt;
+}