+++ test/transform/resource/before/ArgumentOptionalIntFragment.java
@@ -1,0 +1,10 @@
+// ignore
+
+import android.app.Fragment;
+import android.os.Bundle;
+
+public class ArgumentOptionalIntFragment extends Fragment {
+	
+	@hrisey.Argument(optional = true)
+	private int myInt;
+}