+++ test/transform/resource/before/ExtraIntActivity.java
@@ -1,0 +1,10 @@
+// ignore
+
+import android.app.Activity;
+import android.os.Bundle;
+
+public class ExtraIntActivity extends Activity {
+	
+	@hrisey.Extra
+	private int myInt;
+}