+++ test/transform/resource/before/InstanceStateIntFragment.java
@@ -1,0 +1,10 @@
+// ignore
+
+import android.app.Fragment;
+import android.os.Bundle;
+
+public class InstanceStateIntFragment extends Fragment {
+	
+	@hrisey.InstanceState
+	private int myInt;
+}