+++ test/transform/resource/before/InstanceStateIntWithOnCreateActivity.java
@@ -1,0 +1,20 @@
+// ignore
+
+import android.app.Activity;
+import android.os.Bundle;
+
+public class InstanceStateIntWithOnCreateActivity extends Activity {
+	
+	@hrisey.InstanceState
+	private int myInt;
+	
+	@Override
+	protected void onCreate(Bundle state) {
+		super.onCreate(state);
+	}
+	
+	@Override
+	protected void onSaveInstanceState(Bundle state) {
+		super.onSaveInstanceState(state);
+	}
+}