+++ test/transform/resource/before/PreferencesDouble.java
@@ -1,0 +1,4 @@
+@hrisey.Preferences
+class PrefsClass {
+	private double myDouble;
+}