+++ test/transform/resource/before/PreferencesInt.java
@@ -1,0 +1,4 @@
+@hrisey.Preferences
+class PrefsClass {
+	private int myInt;
+}