+++ test/transform/resource/before/PreferencesObject.java
@@ -1,0 +1,11 @@
+import java.util.List;
+
+@hrisey.Preferences
+class PrefsClass {
+	private MyObj myObj;
+}
+
+final class MyObj {
+	int myInt;
+	List<String> myStrings;
+}