+++ test/transform/resource/before/PreferencesObjectGeneric.java
@@ -1,0 +1,6 @@
+import java.util.List;
+
+@hrisey.Preferences
+class PrefsClass {
+	private List<String> myList;
+}