+++ test/transform/resource/before/PreferencesPrimitives.java
@@ -1,0 +1,8 @@
+@hrisey.Preferences
+class PrefsClass {
+	private final boolean myBool = true;
+	private final float myFloat = 0.5F;
+	private final int myInt = 666;
+	private final long myLong = 100000000000000000L;
+	private final String myString = "mg6";
+}