+++ test/transform/resource/before/PreferencesString.java
@@ -1,0 +1,4 @@
+@hrisey.Preferences
+class PrefsClass {
+	private String myString;
+}